+++ src/libs/service/backgroundDAppService.ts
@@ -35,6 +35,7 @@ import {
   tonConnectDisconnect,
   tonConnectRequest,
   tonConnectTransaction,
+  tonReConnectRequest,
 } from "./dApp/tonConnectService";
 import {
   confirmAccountSeqNo,
@@ -227,6 +228,9 @@ const handleDAppMessage = async (message: DAppMessage): Promise<unknown> => {
 
     case "tonConnect_connect": {
       return tonConnectRequest(message.id, origin, message.params[0]);
+    }
+    case "tonConnect_reconnect": {
+      return tonReConnectRequest(origin);
     }
     case "tonConnect_disconnect": {
       return tonConnectDisconnect(message.id, origin);